Remove MetricsMonitor file and unused references
This commit removes the MetricsMonitor.cs file which is no longer needed. It also cleans up several other files by removing references that are not being utilized. This removal of unused imports helps to keep the codebase cleaner and easier to read. Additionally, the version of certain testing dependencies in XFramework.Tests.csproj has been updated which should improve the quality and reliability of testing.